Move RdfItemReader and Writer to separate project
