+++ rdf/src/test/java/com/marklogic/spring/batch/item/rdf/ImportRdfFromFileJob.java
@@ -1,8 +1,6 @@
-package com.marklogic.spring.batch.samples;
+package com.marklogic.spring.batch.item.rdf;
 
 import com.marklogic.client.helper.DatabaseClientProvider;
-import com.marklogic.spring.batch.item.reader.RdfTripleItemReader;
-import com.marklogic.spring.batch.item.writer.RdfTripleItemWriter;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
