+++ rdf/src/test/resources/job.properties
@@ -1,0 +1,7 @@
+marklogic.host=oscar
+marklogic.port=8200
+marklogic.username=admin
+marklogic.password=admin
+
+marklogic.name=marklogic-spring-batch-test-content
+mlJobRepositoryName=marklogic-spring-batch-test-content
