+++ rdf/src/test/resources/triple/test1.ttl
