+++ rdf/src/test/resources/triple/tigers.ttl
