+++ samples/src/test/java/com.marklogic.spring.batch.samples/DeleteDocumentsJobTest.java
@@ -7,6 +7,7 @@ import com.marklogic.client.io.StringHandle;
 import com.marklogic.spring.batch.test.AbstractJobRunnerTest;
 import org.junit.Assume;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.springframework.batch.core.BatchStatus;
 import org.springframework.batch.core.JobExecution;
@@ -33,6 +34,7 @@ public class DeleteDocumentsJobTest extends AbstractJobRunnerTest {
     }
 
     @Test
+    @Ignore
     public void deleteMonsterCollectionWithDmsdkTest() throws Exception {
         Assume.assumeTrue(isMarkLogic9());
         JobParametersBuilder jpb = new JobParametersBuilder();
