+++ settings.gradle
@@ -2,3 +2,4 @@ include "core"
 include "infrastructure"
 include "test"
 include "samples"
+include "rdf"
